packages/server/requirements.txt
# FastAPI Server for PyIDE Remote Kernel
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
websockets==12.0
pydantic-settings==2.1.0
docker==7.0.0
python-multipart==0.0.9
psutil>=5.9.0
